[syscall] Remove syzkaller from manifest
syzkaller is no longer included in the build locally, only via a build
elsewhere that uses a build of it that's not part of the main Fuchsia
tree. Remove this dependency for now to avoid confusion (and because the
current for has uses of zx_task_resume() which we're in the process of
nuking.

ZX-2720 #comment

Test: CQ
Change-Id: I8e3020192eff15c98aa00c499f7c9b26f782a3b3